requirements/base.txt
django==4.2.7
psycopg2-binary==2.9.9
python-environ==0.4.54
django-cors-headers==4.3.1
django-jazzmin==2.6.0
drf-yasg==1.21.7
djangorestframework==3.15.1
django-filter==23.4
django-rosetta==0.9.9
python-telegram-bot==13.15
pillow==10.1.0
django_celery_beat==2.6.0
django-debug-toolbar==4.3.0
setuptools==68.2.0
redis==5.0.3
